ecommerce_project/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
defusedxml==0.7.1
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-xml==2.0.0
idna==3.10
mysqlclient==2.2.7
oauthlib==3.2.2
pillow==11.1.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
